Switched from `f64` to `Decimal` for lat/lng coordinates.

+++ src/geocoding/response/geometry.rs
@@ -7,6 +7,7 @@ use crate::{
     geocoding::location_type::LocationType,
     latlng::LatLng,
 }; // use
+use rust_decimal::Decimal;
 use serde::{Serialize, Deserialize};
 
 /// Contains the geocoded latitude/longitude, recommended viewport for
@@ -48,7 +49,7 @@ impl Geometry {
     /// ```rust
     /// let bounds_southwest_lat = geocoding.geometry.get_bounds_southwest_lng();
     /// ```
-    pub fn get_bounds_southwest_lat(&self) -> Option<f64> {
+    pub fn get_bounds_southwest_lat(&self) -> Option<Decimal> {
         match &self.bounds {
             Some(bounds) => Some(bounds.southwest.lat),
             None => None,
@@ -62,7 +63,7 @@ impl Geometry {
     /// ```rust
     /// let bounds_southwest_lng = geocoding.geometry.get_bounds_southwest_lng();
     /// ```
-    pub fn get_bounds_southwest_lng(&self) -> Option<f64> {
+    pub fn get_bounds_southwest_lng(&self) -> Option<Decimal> {
         match &self.bounds {
             Some(bounds) => Some(bounds.southwest.lng),
             None => None,
@@ -76,7 +77,7 @@ impl Geometry {
     /// ```rust
     /// let bounds_northeast_lat = geocoding.geometry.get_bounds_northeast_lng();
     /// ```
-    pub fn get_bounds_northeast_lat(&self) -> Option<f64> {
+    pub fn get_bounds_northeast_lat(&self) -> Option<Decimal> {
         match &self.bounds {
             Some(bounds) => Some(bounds.northeast.lat),
             None => None,
@@ -90,7 +91,7 @@ impl Geometry {
     /// ```rust
     /// let bounds_northeast_lng = geocoding.geometry.get_bounds_northeast_lng();
     /// ```
-    pub fn get_bounds_northeast_lng(&self) -> Option<f64> {
+    pub fn get_bounds_northeast_lng(&self) -> Option<Decimal> {
         match &self.bounds {
             Some(bounds) => Some(bounds.northeast.lng),
             None => None,
